# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API Documentation Portal
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface Group {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  icon: string;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface GroupInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  icon: string;
}

export interface GroupUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  icon?: string;
  sortOrder?: number;
}

export type EndpointMethod = typeof EndpointMethod[keyof typeof EndpointMethod];


export const EndpointMethod = {
  GET: 'GET',
  POST: 'POST',
  PUT: 'PUT',
  PATCH: 'PATCH',
  DELETE: 'DELETE',
} as const;

export type EndpointStatus = typeof EndpointStatus[keyof typeof EndpointStatus];


export const EndpointStatus = {
  published: 'published',
  draft: 'draft',
  deprecated: 'deprecated',
} as const;

export interface Endpoint {
  id: number;
  groupId: number;
  method: EndpointMethod;
  path: string;
  summary: string;
  /** @nullable */
  description?: string | null;
  status: EndpointStatus;
  /**
     * JSON string of parameter objects
     * @nullable
     */
  params?: string | null;
  /**
     * JSON string of example response
     * @nullable
     */
  responseExample?: string | null;
  /** @nullable */
  responseStatus?: number | null;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export type EndpointInputMethod = typeof EndpointInputMethod[keyof typeof EndpointInputMethod];


export const EndpointInputMethod = {
  GET: 'GET',
  POST: 'POST',
  PUT: 'PUT',
  PATCH: 'PATCH',
  DELETE: 'DELETE',
} as const;

export type EndpointInputStatus = typeof EndpointInputStatus[keyof typeof EndpointInputStatus];


export const EndpointInputStatus = {
  published: 'published',
  draft: 'draft',
  deprecated: 'deprecated',
} as const;

export interface EndpointInput {
  groupId: number;
  method: EndpointInputMethod;
  /** @minLength 1 */
  path: string;
  /** @minLength 1 */
  summary: string;
  description?: string;
  status?: EndpointInputStatus;
  /** JSON string of parameter objects */
  params?: string;
  responseExample?: string;
  responseStatus?: number;
}

export type EndpointUpdateMethod = typeof EndpointUpdateMethod[keyof typeof EndpointUpdateMethod];


export const EndpointUpdateMethod = {
  GET: 'GET',
  POST: 'POST',
  PUT: 'PUT',
  PATCH: 'PATCH',
  DELETE: 'DELETE',
} as const;

export type EndpointUpdateStatus = typeof EndpointUpdateStatus[keyof typeof EndpointUpdateStatus];


export const EndpointUpdateStatus = {
  published: 'published',
  draft: 'draft',
  deprecated: 'deprecated',
} as const;

export interface EndpointUpdate {
  groupId?: number;
  method?: EndpointUpdateMethod;
  /** @minLength 1 */
  path?: string;
  /** @minLength 1 */
  summary?: string;
  description?: string;
  status?: EndpointUpdateStatus;
  params?: string;
  responseExample?: string;
  responseStatus?: number;
}

export interface Stats {
  total: number;
  published: number;
  draft: number;
  deprecated: number;
  groups: number;
}

export type ListEndpointsParams = {
groupId?: number;
status?: ListEndpointsStatus;
q?: string;
};

export type ListEndpointsStatus = typeof ListEndpointsStatus[keyof typeof ListEndpointsStatus];


export const ListEndpointsStatus = {
  published: 'published',
  draft: 'draft',
  deprecated: 'deprecated',
} as const;