# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API Documentation Portal
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: groups
    description: Endpoint group operations
  - name: endpoints
    description: API endpoint operations
  - name: stats
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List all endpoint groups
      description: Returns all groups ordered by sort_order
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new endpoint group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"

  /groups/{id}:
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update an endpoint group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /endpoints:
    get:
      operationId: listEndpoints
      tags: [endpoints]
      summary: List all API endpoints
      description: Returns endpoints, optionally filtered by group, status, or search query
      parameters:
        - name: groupId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [published, draft, deprecated]
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of endpoints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Endpoint"
    post:
      operationId: createEndpoint
      tags: [endpoints]
      summary: Create a new API endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndpointInput"
      responses:
        "201":
          description: Created endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Endpoint"

  /endpoints/{id}:
    get:
      operationId: getEndpoint
      tags: [endpoints]
      summary: Get a single endpoint by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Endpoint detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Endpoint"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEndpoint
      tags: [endpoints]
      summary: Update an API endpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndpointUpdate"
      responses:
        "200":
          description: Updated endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Endpoint"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEndpoint
      tags: [endpoints]
      summary: Delete an API endpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get dashboard statistics
      description: Returns aggregate counts — total endpoints, published, draft, deprecated, and group count
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    Group:
      type: object
      required: [id, name, icon, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    GroupInput:
      type: object
      required: [name, icon]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        icon:
          type: string

    GroupUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer

    Endpoint:
      type: object
      required: [id, groupId, method, path, summary, status, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: integer
        groupId:
          type: integer
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        path:
          type: string
        summary:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [published, draft, deprecated]
        params:
          type: ["string", "null"]
          description: JSON string of parameter objects
        responseExample:
          type: ["string", "null"]
          description: JSON string of example response
        responseStatus:
          type: ["integer", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    EndpointInput:
      type: object
      required: [groupId, method, path, summary]
      properties:
        groupId:
          type: integer
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        path:
          type: string
          minLength: 1
        summary:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [published, draft, deprecated]
        params:
          type: string
          description: JSON string of parameter objects
        responseExample:
          type: string
        responseStatus:
          type: integer

    EndpointUpdate:
      type: object
      properties:
        groupId:
          type: integer
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        path:
          type: string
          minLength: 1
        summary:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [published, draft, deprecated]
        params:
          type: string
        responseExample:
          type: string
        responseStatus:
          type: integer

    Stats:
      type: object
      required: [total, published, draft, deprecated, groups]
      properties:
        total:
          type: integer
        published:
          type: integer
        draft:
          type: integer
        deprecated:
          type: integer
        groups:
          type: integer