# lib/api-zod/src/generated/types/endpoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API Documentation Portal
 * OpenAPI spec version: 0.1.0
 */
import type { EndpointMethod } from './endpointMethod';
import type { EndpointStatus } from './endpointStatus';

export interface Endpoint {
  id: number;
  groupId: number;
  method: EndpointMethod;
  path: string;
  summary: string;
  /** @nullable */
  description?: string | null;
  status: EndpointStatus;
  /**
     * JSON string of parameter objects
     * @nullable
     */
  params?: string | null;
  /**
     * JSON string of example response
     * @nullable
     */
  responseExample?: string | null;
  /** @nullable */
  responseStatus?: number | null;
  sortOrder: number;
  createdAt: Date;
  updatedAt: Date;
}