# lib/api-zod/src/generated/types/endpointInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API Documentation Portal
 * OpenAPI spec version: 0.1.0
 */
import type { EndpointInputMethod } from './endpointInputMethod';
import type { EndpointInputStatus } from './endpointInputStatus';

export interface EndpointInput {
  groupId: number;
  method: EndpointInputMethod;
  /** @minLength 1 */
  path: string;
  /** @minLength 1 */
  summary: string;
  description?: string;
  status?: EndpointInputStatus;
  /** JSON string of parameter objects */
  params?: string;
  responseExample?: string;
  responseStatus?: number;
}