# lib/api-zod/src/generated/types/endpointUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API Documentation Portal
 * OpenAPI spec version: 0.1.0
 */
import type { EndpointUpdateMethod } from './endpointUpdateMethod';
import type { EndpointUpdateStatus } from './endpointUpdateStatus';

export interface EndpointUpdate {
  groupId?: number;
  method?: EndpointUpdateMethod;
  /** @minLength 1 */
  path?: string;
  /** @minLength 1 */
  summary?: string;
  description?: string;
  status?: EndpointUpdateStatus;
  params?: string;
  responseExample?: string;
  responseStatus?: number;
}