# lib/api-zod/src/generated/types/group.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API Documentation Portal
 * OpenAPI spec version: 0.1.0
 */

export interface Group {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  icon: string;
  sortOrder: number;
  createdAt: Date;
  updatedAt: Date;
}