# lib/api-zod/src/generated/types/groupInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API Documentation Portal
 * OpenAPI spec version: 0.1.0
 */

export interface GroupInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  icon: string;
}